# Use Go 1.17 for fixed size copying