Experiment. Checking if worth it.

```
benchmark                                          old ns/op     new ns/op     delta 
BenchmarkDecompress4XTable/digits-32               171211        163020        -4.78%
BenchmarkDecompress4XTable/gettysburg-32           4045          3938          -2.65%
BenchmarkDecompress4XTable/twain-32                585484        567582        -3.06%
BenchmarkDecompress4XTable/low-ent.10k-32          57526         57130         -0.69%
BenchmarkDecompress4XTable/superlow-ent-10k-32     15864         15675         -1.19%
BenchmarkDecompress4XTable/case1-32                2003          2023          +1.00%
BenchmarkDecompress4XTable/case2-32                1993          1982          -0.55%
BenchmarkDecompress4XTable/case3-32                1987          2026          +1.96%
BenchmarkDecompress4XTable/pngdata.001-32          82474         78583         -4.72%
BenchmarkDecompress4XTable/normcount2-32           1377          1329          -3.49%
                                                                                     
benchmark                                          old MB/s     new MB/s     speedup 
BenchmarkDecompress4XTable/digits-32               584.09       613.44       1.05x   
BenchmarkDecompress4XTable/gettysburg-32           382.67       393.10       1.03x   
BenchmarkDecompress4XTable/twain-32                447.74       461.86       1.03x   
BenchmarkDecompress4XTable/low-ent.10k-32          695.34       700.16       1.01x   
BenchmarkDecompress4XTable/superlow-ent-10k-32     661.86       669.84       1.01x   
BenchmarkDecompress4XTable/case1-32                27.46        27.18        0.99x   
BenchmarkDecompress4XTable/case2-32                22.58        22.70        1.01x   
BenchmarkDecompress4XTable/case3-32                24.16        23.69        0.98x   
BenchmarkDecompress4XTable/pngdata.001-32          620.80       651.54       1.05x   
BenchmarkDecompress4XTable/normcount2-32           63.18        65.48        1.04x   
```